Proving System API changes
